Module k8s_cri::v1alpha2

API version v1alpha2, original Protocol Buffers file.

Modules

image_service_client

Generated client implementations.

image_service_server

Generated server implementations.

runtime_service_client

Generated client implementations.

runtime_service_server

Generated server implementations.

security_profile

Nested message and enum types in SecurityProfile.

Structs

AttachRequest

AttachResponse

AuthConfig

AuthConfig contains authorization information for connecting to a registry.

Capability

Capability contains the container capabilities to add or drop Dropping a capability will drop it from all sets. If a capability is added to only the add_capabilities list then it gets added to permitted, inheritable, effective and bounding sets, i.e. all sets except the ambient set. If a capability is added to only the add_ambient_capabilities list then it gets added to all sets, i.e permitted inheritable, effective, bounding and ambient sets. If a capability is added to add_capabilities and add_ambient_capabilities lists then it gets added to all sets, i.e. permitted, inheritable, effective, bounding and ambient sets.

Container

Container provides the runtime information for a container, such as ID, hash, state of the container.

ContainerAttributes

ContainerAttributes provides basic information of the container.

ContainerConfig

ContainerConfig holds all the required and optional fields for creating a container.

ContainerFilter

ContainerFilter is used to filter containers. All those fields are combined with ‘AND’

ContainerMetadata

ContainerMetadata holds all necessary information for building the container name. The container runtime is encouraged to expose the metadata in its user interface for better user experience. E.g., runtime can construct a unique container name based on the metadata. Note that (name, attempt) is unique within a sandbox for the entire lifetime of the sandbox.

ContainerResources

ContainerResources holds resource limits configuration for a container.

ContainerStateValue

ContainerStateValue is the wrapper of ContainerState.

ContainerStats

ContainerStats provides the resource usage statistics for a container.

ContainerStatsFilter

ContainerStatsFilter is used to filter containers. All those fields are combined with ‘AND’

ContainerStatsRequest

ContainerStatsResponse

ContainerStatus

ContainerStatus represents the status of a container.

ContainerStatusRequest

ContainerStatusResponse

CpuUsage

CpuUsage provides the CPU usage information.

CreateContainerRequest

CreateContainerResponse

Device

Device specifies a host device to mount into a container.

DnsConfig

DNSConfig specifies the DNS servers and search domains of a sandbox.

ExecRequest

ExecResponse

ExecSyncRequest

ExecSyncResponse

FilesystemIdentifier

FilesystemIdentifier uniquely identify the filesystem.

FilesystemUsage

FilesystemUsage provides the filesystem usage information.

HugepageLimit

HugepageLimit corresponds to the filehugetlb.<hugepagesize>.limit_in_byte in container level cgroup. For example, PageSize=1GB, Limit=1073741824 means setting 1073741824 bytes to hugetlb.1GB.limit_in_bytes.

IdMapping

IDMapping describes host to container ID mappings for a pod sandbox.

Image

Basic information about a container image.

ImageFilter

ImageFsInfoRequest

ImageFsInfoResponse

ImageSpec

ImageSpec is an internal representation of an image.

ImageStatusRequest

ImageStatusResponse

Int64Value

Int64Value is the wrapper of int64.

KeyValue

LinuxContainerConfig

LinuxContainerConfig contains platform-specific configuration for Linux-based containers.

LinuxContainerResources

LinuxContainerResources specifies Linux specific configuration for resources.

LinuxContainerSecurityContext

LinuxContainerSecurityContext holds linux security configuration that will be applied to a container.

LinuxPodSandboxConfig

LinuxPodSandboxConfig holds platform-specific configurations for Linux host platforms and Linux-based containers.

LinuxPodSandboxStats

LinuxPodSandboxStats provides the resource usage statistics for a pod sandbox on linux.

LinuxPodSandboxStatus

LinuxSandboxStatus contains status specific to Linux sandboxes.

LinuxSandboxSecurityContext

LinuxSandboxSecurityContext holds linux security configuration that will be applied to a sandbox. Note that:

ListContainerStatsRequest

ListContainerStatsResponse

ListContainersRequest

ListContainersResponse

ListImagesRequest

ListImagesResponse

ListPodSandboxRequest

ListPodSandboxResponse

ListPodSandboxStatsRequest

ListPodSandboxStatsResponse

MemoryUsage

MemoryUsage provides the memory usage information.

Mount

Mount specifies a host volume to mount into a container.

Namespace

Namespace contains paths to the namespaces.

NamespaceOption

NamespaceOption provides options for Linux namespaces.

NetworkConfig

NetworkInterfaceUsage

NetworkInterfaceUsage contains resource value data about a network interface.

NetworkUsage

NetworkUsage contains data about network resources.

PodIp

PodIP represents an ip of a Pod

PodSandbox

PodSandbox contains minimal information about a sandbox.

PodSandboxAttributes

PodSandboxAttributes provides basic information of the pod sandbox.

PodSandboxConfig

PodSandboxConfig holds all the required and optional fields for creating a sandbox.

PodSandboxFilter

PodSandboxFilter is used to filter a list of PodSandboxes. All those fields are combined with ‘AND’

PodSandboxMetadata

PodSandboxMetadata holds all necessary information for building the sandbox name. The container runtime is encouraged to expose the metadata associated with the PodSandbox in its user interface for better user experience. For example, the runtime can construct a unique PodSandboxName based on the metadata.

PodSandboxNetworkStatus

PodSandboxNetworkStatus is the status of the network for a PodSandbox. Currently ignored for pods sharing the host networking namespace.

PodSandboxStateValue

PodSandboxStateValue is the wrapper of PodSandboxState.

PodSandboxStats

PodSandboxStats provides the resource usage statistics for a pod. The linux or windows field will be populated depending on the platform.

PodSandboxStatsFilter

PodSandboxStatsFilter is used to filter the list of pod sandboxes to retrieve stats for. All those fields are combined with ‘AND’.

PodSandboxStatsRequest

PodSandboxStatsResponse

PodSandboxStatus

PodSandboxStatus contains the status of the PodSandbox.

PodSandboxStatusRequest

PodSandboxStatusResponse

PortForwardRequest

PortForwardResponse

PortMapping

PortMapping specifies the port mapping configurations of a sandbox.

ProcessUsage

ProcessUsage are stats pertaining to processes.

PullImageRequest

PullImageResponse

RemoveContainerRequest

RemoveContainerResponse

RemoveImageRequest

RemoveImageResponse

RemovePodSandboxRequest

RemovePodSandboxResponse

ReopenContainerLogRequest

ReopenContainerLogResponse

RunPodSandboxRequest

RunPodSandboxResponse

RuntimeCondition

RuntimeCondition contains condition information for the runtime. There are 2 kinds of runtime conditions:

RuntimeConfig

RuntimeStatus

RuntimeStatus is information about the current status of the runtime.

SeLinuxOption

SELinuxOption are the labels to be applied to the container.

SecurityProfile

A security profile which can be used for sandboxes and containers.

StartContainerRequest

StartContainerResponse

StatusRequest

StatusResponse

StopContainerRequest

StopContainerResponse

StopPodSandboxRequest

StopPodSandboxResponse

UInt64Value

UInt64Value is the wrapper of uint64.

UpdateContainerResourcesRequest

UpdateContainerResourcesResponse

UpdateRuntimeConfigRequest

UpdateRuntimeConfigResponse

UserNamespace

UserNamespace describes the intended user namespace configuration for a pod sandbox.

VersionRequest

VersionResponse

WindowsContainerConfig

WindowsContainerConfig contains platform-specific configuration for Windows-based containers.

WindowsContainerResources

WindowsContainerResources specifies Windows specific configuration for resources.

WindowsContainerSecurityContext

WindowsContainerSecurityContext holds windows security configuration that will be applied to a container.

WindowsPodSandboxConfig

WindowsPodSandboxConfig holds platform-specific configurations for Windows host platforms and Windows-based containers.

WindowsPodSandboxStats

WindowsPodSandboxStats provides the resource usage statistics for a pod sandbox on windows

WindowsSandboxSecurityContext

WindowsSandboxSecurityContext holds platform-specific configurations that will be applied to a sandbox. These settings will only apply to the sandbox container.

Enums

ContainerState

MountPropagation

NamespaceMode

A NamespaceMode describes the intended namespace configuration for each of the namespaces (Network, PID, IPC) in NamespaceOption. Runtimes should map these modes as appropriate for the technology underlying the runtime.

PodSandboxState

Protocol