Trait k8s_cri::v1alpha2::image_service_server::ImageService

pub trait ImageService: Send + Sync + 'static {
    // Required methods
    fn list_images<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListImagesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ListImagesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn image_status<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ImageStatusRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ImageStatusResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pull_image<'life0, 'async_trait>(
        &'life0 self,
        request: Request<PullImageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<PullImageResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove_image<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RemoveImageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<RemoveImageResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn image_fs_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ImageFsInfoRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ImageFsInfoResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ImageServiceServer.

Required Methods

fn list_images<'life0, 'async_trait>( &'life0 self, request: Request<ListImagesRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<ListImagesResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

ListImages lists existing images.

fn image_status<'life0, 'async_trait>( &'life0 self, request: Request<ImageStatusRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<ImageStatusResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

ImageStatus returns the status of the image. If the image is not present, returns a response with ImageStatusResponse.Image set to nil.

fn pull_image<'life0, 'async_trait>( &'life0 self, request: Request<PullImageRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<PullImageResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

PullImage pulls an image with authentication config.

fn remove_image<'life0, 'async_trait>( &'life0 self, request: Request<RemoveImageRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<RemoveImageResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

RemoveImage removes the image. This call is idempotent, and must not return an error if the image has already been removed.

fn image_fs_info<'life0, 'async_trait>( &'life0 self, request: Request<ImageFsInfoRequest> ) -> Pin<Box<dyn Future<Output = Result<Response<ImageFsInfoResponse>, Status>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

ImageFSInfo returns information of the filesystem that is used to store images.