Struct k8s_cri::v1alpha2::LinuxContainerSecurityContext

pub struct LinuxContainerSecurityContext {

    pub capabilities: Option<Capability>,
    pub privileged: bool,
    pub namespace_options: Option<NamespaceOption>,
    pub selinux_options: Option<SeLinuxOption>,
    pub run_as_user: Option<Int64Value>,
    pub run_as_group: Option<Int64Value>,
    pub run_as_username: String,
    pub readonly_rootfs: bool,
    pub supplemental_groups: Vec<i64>,
    pub no_new_privs: bool,
    pub masked_paths: Vec<String>,
    pub readonly_paths: Vec<String>,
    pub seccomp: Option<SecurityProfile>,
    pub apparmor: Option<SecurityProfile>,
    pub apparmor_profile: String,
    pub seccomp_profile_path: String,

}

LinuxContainerSecurityContext holds linux security configuration that will be applied to a container.

Fields
capabilities: Option<Capability>

Capabilities to add or drop.

privileged: bool

If set, run container in privileged mode. Privileged mode is incompatible with the following options. If privileged is set, the following features MAY have no effect:

capabilities

selinux_options

seccomp

apparmor

Privileged mode implies the following specific options are applied:

All capabilities are added.

Sensitive paths, such as kernel module paths within sysfs, are not masked.

Any sysfs and procfs mounts are mounted RW.

AppArmor confinement is not applied.

Seccomp restrictions are not applied.

The device cgroup does not restrict access to any devices.

All devices from the hostโs /dev are available within the container.

SELinux restrictions are not applied (e.g. label=disabled).

namespace_options: Option<NamespaceOption>

Configurations for the containerโs namespaces. Only used if the container uses namespace for isolation.

selinux_options: Option<SeLinuxOption>

SELinux context to be optionally applied.

run_as_user: Option<Int64Value>

UID to run the container process as. Only one of run_as_user and run_as_username can be specified at a time.

run_as_group: Option<Int64Value>

GID to run the container process as. run_as_group should only be specified when run_as_user or run_as_username is specified; otherwise, the runtime MUST error.

run_as_username: String

User name to run the container process as. If specified, the user MUST exist in the container image (i.e. in the /etc/passwd inside the image), and be resolved there by the runtime; otherwise, the runtime MUST error.

readonly_rootfs: bool

If set, the root filesystem of the container is read-only.

supplemental_groups: Vec<i64>

List of groups applied to the first process run in the container, in addition to the containerโs primary GID.

no_new_privs: bool

no_new_privs defines if the flag for no_new_privs should be set on the container.

masked_paths: Vec<String>

masked_paths is a slice of paths that should be masked by the container runtime, this can be passed directly to the OCI spec.

readonly_paths: Vec<String>

readonly_paths is a slice of paths that should be set as readonly by the container runtime, this can be passed directly to the OCI spec.

seccomp: Option<SecurityProfile>

Seccomp profile for the container.

apparmor: Option<SecurityProfile>

AppArmor profile for the container.

apparmor_profile: String
๐Deprecated

AppArmor profile for the container, candidate values are:

runtime/default: equivalent to not specifying a profile.

unconfined: no profiles are loaded

localhost/<profile_name>: profile loaded on the node (localhost) by name. The possible profile names are detailed at https://gitlab.com/apparmor/apparmor/-/wikis/AppArmor_Core_Policy_Reference

seccomp_profile_path: String
๐Deprecated

Seccomp profile for the container, candidate values are:

runtime/default: the default profile for the container runtime

unconfined: unconfined profile, ie, no seccomp sandboxing

localhost/: the profile installed on the node. is the full path of the profile. Default: โโ, which is identical with unconfined.